pyproject.toml
[project]
name = "helper-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.20",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "langgraph-checkpoint-sqlite>=2.0.6",
    "langsmith>=0.3.18",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "gradio",
    "beautifulsoup4",
    "playwright"
]
[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
